# maint(index) Upgrade tantivy 0.22 -> 0.24