--- a/core/src/rust/filodb_core/Cargo.toml
+++ b/core/src/rust/filodb_core/Cargo.toml
@@ -18,8 +18,8 @@ num-derive = "0.4.2"
 num-traits = "0.2.19"
 quick_cache = { version = "0.6.2", features = ["stats"] }
 regex = "1.10.5"
-tantivy = "0.22.0"
-tantivy-common = "0.7.0"
+tantivy = "0.24.1"
+tantivy-common = "0.9.0"
 tantivy-fst = "0.5.0"
 tantivy_utils = { path = "../tantivy_utils" }
 thiserror = "1.0.62"