--- a/core/src/rust/filodb_core/src/query_parser.rs
+++ b/core/src/rust/filodb_core/src/query_parser.rs
@@ -298,10 +298,11 @@ fn parse_long_range_query<'a>(
     // 8 byte end
     let (input, end) = le_i64(input)?;
 
-    let query = RangeQuery::new_i64_bounds(
-        field_name.to_string(),
-        Bound::Included(start),
-        Bound::Included(end),
+    let field = schema.get_field(field_name).to_nom_err()?;
+
+    let query = RangeQuery::new(
+        Bound::Included(Term::from_field_i64(field, start)),
+        Bound::Included(Term::from_field_i64(field, end)),
     );
 
     Ok((input, Box::new(query)))