--- a/core/src/rust/tantivy_utils/Cargo.toml
+++ b/core/src/rust/tantivy_utils/Cargo.toml
@@ -5,9 +5,9 @@ edition = "2021"
 rust-version = "1.78"
 
 [dependencies]
-hashbrown = "0.14.5"
+hashbrown = "0.15.2"
 nohash-hasher = "0.2.0"
 quick_cache = { version = "0.6.2", features = ["stats"] }
-tantivy = "0.22.0"
-tantivy-common = "0.7.0"
+tantivy = "0.24.1"
+tantivy-common = "0.9.0"
 tantivy-fst = "0.5.0"