--- a/core/src/rust/tantivy_utils/src/collectors/column_cache.rs
+++ b/core/src/rust/tantivy_utils/src/collectors/column_cache.rs
@@ -3,9 +3,10 @@
 use std::sync::Arc;
 
 use quick_cache::{sync::Cache, Equivalent};
+use tantivy::index::SegmentId;
 use tantivy::{
     columnar::{BytesColumn, Column, DynamicColumn, HasAssociatedColumnType, StrColumn},
-    SegmentId, SegmentReader,
+    SegmentReader,
 };
 
 // Max column items to cache.  These are relatively cheap (< 1KB)