--- a/core/src/rust/tantivy_utils/src/collectors/part_key_collector.rs
+++ b/core/src/rust/tantivy_utils/src/collectors/part_key_collector.rs
@@ -21,9 +21,10 @@ impl UnresolvedPartKey {
         let part_key_field = searcher.schema().get_field(PART_KEY)?;
 
         let Some(OwnedValue::Bytes(part_key)) = doc_data
+            .field_values()
             .into_iter()
-            .filter(|x| x.field == part_key_field)
-            .map(|x| x.value)
+            .filter(|(field, _)| *field == part_key_field)
+            .map(|(_, value)| value.into())
             .next()
         else {
             return Err(TantivyError::FieldNotFound(PART_KEY.to_string()));