--- a/core/src/rust/tantivy_utils/src/collectors/part_key_collector.rs
+++ b/core/src/rust/tantivy_utils/src/collectors/part_key_collector.rs
@@ -22,7 +22,6 @@ impl UnresolvedPartKey {
 
         let Some(OwnedValue::Bytes(part_key)) = doc_data
             .field_values()
-            .into_iter()
             .filter(|(field, _)| *field == part_key_field)
             .map(|(_, value)| value.into())
             .next()