**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [ ] Docs have been added / updated (for bug fixes / features) ?

**Current behavior :** 
No change in behavior expected, tantity upgraded

**New behavior :**
NA


Following are the results for the JMH tests

Lucene 
```
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                         Mode  Cnt        Score       Error  Units
[info] PartKeyLuceneIndexBenchmark.emptyPartIdsLookupWithEqualsFilters  thrpt    5    59244.289 ±  3745.757  ops/s
[info] PartKeyLuceneIndexBenchmark.getLabelNames                        thrpt    5      496.276 ±    10.932  ops/s
[info] PartKeyLuceneIndexBenchmark.getMetricNames                       thrpt    5      417.537 ±     5.399  ops/s
[info] PartKeyLuceneIndexBenchmark.getMetricNamesSearchAndIterate       thrpt    5        0.595 ±     0.029  ops/s
[info] PartKeyLuceneIndexBenchmark.indexNames                           thrpt    5  2669633.267 ± 63241.077  ops/s
[info] PartKeyLuceneIndexBenchmark.indexValues                          thrpt    5     2364.378 ±    80.261  ops/s
[info] PartKeyLuceneIndexBenchmark.partIdsLookupOverTime                thrpt    5     1108.198 ±    28.733  ops/s
[info] PartKeyLuceneIndexBenchmark.partIdsLookupWithEnumRegexFilter     thrpt    5     9157.707 ±   159.633  ops/s
[info] PartKeyLuceneIndexBenchmark.partIdsLookupWithEqualsFilters       thrpt    5      119.867 ±     3.743  ops/s
[info] PartKeyLuceneIndexBenchmark.partIdsLookupWithPrefixRegexFilters  thrpt    5     2024.598 ±   100.499  ops/s
[info] PartKeyLuceneIndexBenchmark.partIdsLookupWithSuffixRegexFilters  thrpt    5     1285.420 ±    62.553  ops/s
[info] PartKeyLuceneIndexBenchmark.startTimeLookupWithPartId            thrpt    5     8730.719 ±   212.746  ops/s
[info] Benchmark result is saved to jmh-result.json
```

Tantivy 0.24 
```
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                          Mode  Cnt       Score      Error  Units
[info] PartKeyTantivyIndexBenchmark.emptyPartIdsLookupWithEqualsFilters  thrpt    5  127239.818 ± 4062.705  ops/s
[info] PartKeyTantivyIndexBenchmark.getLabelNames                        thrpt    5     155.028 ±    1.264  ops/s
[info] PartKeyTantivyIndexBenchmark.getMetricNames                       thrpt    5     660.219 ±   18.836  ops/s
[info] PartKeyTantivyIndexBenchmark.getMetricNamesSearchAndIterate       thrpt    5       0.534 ±    0.071  ops/s
[info] PartKeyTantivyIndexBenchmark.indexNames                           thrpt    5   91897.068 ± 4066.508  ops/s
[info] PartKeyTantivyIndexBenchmark.indexValues                          thrpt    5     964.292 ±   36.884  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupOverTime                thrpt    5   15044.028 ±  678.688  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithEnumRegexFilter     thrpt    5   35119.814 ±  936.187  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithEqualsFilters       thrpt    5     293.185 ±   18.222  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithPrefixRegexFilters  thrpt    5    5553.633 ±  389.551  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithSuffixRegexFilters  thrpt    5    5674.180 ±  591.361  ops/s
[info] PartKeyTantivyIndexBenchmark.startTimeLookupWithPartId            thrpt    5   26556.422 ±  219.451  ops/s
```

Tantivy 0.22
```
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                          Mode  Cnt       Score      Error  Units
[info] PartKeyTantivyIndexBenchmark.emptyPartIdsLookupWithEqualsFilters  thrpt    5  124273.956 ± 8756.878  ops/s
[info] PartKeyTantivyIndexBenchmark.getLabelNames                        thrpt    5     177.894 ±   11.655  ops/s
[info] PartKeyTantivyIndexBenchmark.getMetricNames                       thrpt    5     602.656 ±   20.541  ops/s
[info] PartKeyTantivyIndexBenchmark.getMetricNamesSearchAndIterate       thrpt    5       0.545 ±    0.054  ops/s
[info] PartKeyTantivyIndexBenchmark.indexNames                           thrpt    5   97679.311 ± 2820.490  ops/s
[info] PartKeyTantivyIndexBenchmark.indexValues                          thrpt    5     863.077 ±  145.458  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupOverTime                thrpt    5   15367.359 ±  318.019  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithEnumRegexFilter     thrpt    5   37945.732 ± 2855.462  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithEqualsFilters       thrpt    5     263.394 ±   18.259  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithPrefixRegexFilters  thrpt    5    5006.791 ±  208.549  ops/s
[info] PartKeyTantivyIndexBenchmark.partIdsLookupWithSuffixRegexFilters  thrpt    5    5058.430 ±  276.161  ops/s
[info] PartKeyTantivyIndexBenchmark.startTimeLookupWithPartId            thrpt    5   26160.238 ±  668.057  ops/s
[info] Benchmark result is saved to jmh-result.json
```